requirements.txt
langchain==0.1.7
langchain-community==0.0.20
langchain-openai==0.0.6
pandas==2.2.0
python-dotenv==1.0.0
openai==1.12.0
langchain-google-genai==1.0.1
langchain-anthropic
